+++ packages/app/src/contexts/toast-context.tsx
@@ -1,15 +1,8 @@
-import { createContext, useContext, type ReactNode } from "react";
-import { ToastViewport, useToastHost, type ToastApi } from "@/components/toast-host";
+import type { ReactNode } from "react";
+import { ToastViewport, useToastHost } from "@/components/toast-host";
+import { ToastApiProvider } from "./toast-api-context";
 
-const ToastContext = createContext<ToastApi | null>(null);
-
-export function useToast(): ToastApi {
-  const value = useContext(ToastContext);
-  if (!value) {
-    throw new Error("useToast must be used within ToastProvider");
-  }
-  return value;
-}
+export { ToastApiProvider, useToast } from "./toast-api-context";
 
 export function ToastProvider({ children }: { children: ReactNode }) {
   const { api, toast, dismiss } = useToastHost();
@@ -20,8 +13,4 @@ export function ToastProvider({ children }: { children: ReactNode }) {
       <ToastViewport toast={toast} onDismiss={dismiss} />
     </ToastApiProvider>
   );
-}
-
-export function ToastApiProvider({ api, children }: { api: ToastApi; children: ReactNode }) {
-  return <ToastContext.Provider value={api}>{children}</ToastContext.Provider>;
 }
