+++ packages/app/src/plugins/react-native/modal.tsx
@@ -2,8 +2,7 @@ import { QueryClientProvider, useQueryClient } from "@tanstack/react-query";
 import type { ModalComponent, ModalContentProps, ModalProps } from "@getpaseo/plugin/react-native";
 import { usePluginRuntimeContextBridge } from "@getpaseo/plugin/host";
 import { createContext, useCallback, useContext, useMemo, useRef, type ReactNode } from "react";
-import { AdaptiveModalSheet } from "@/components/adaptive-modal-sheet";
-import { ToastApiProvider, useToast as useAppToast } from "@/contexts/toast-context";
+import { ToastApiProvider, useToast as useAppToast } from "@/contexts/toast-api-context";
 
 interface ModalContextValue {
   open: boolean;
@@ -40,6 +39,8 @@ function ModalContent({ title, children }: ModalContentProps) {
     ),
     [bridgePluginRuntime, queryClient, toast],
   );
+  const { AdaptiveModalSheet } =
+    require("../../components/adaptive-modal-sheet") as typeof import("../../components/adaptive-modal-sheet");
 
   return (
     <AdaptiveModalSheet
